+++ front/package.json
@@ -7,6 +7,8 @@
     "@material-ui/core": "^4.7.1",
     "@material-ui/icons": "^4.5.1",
     "@types/jest": "^24.0.24",
+    "@types/node": "^13.1.4",
+    "@types/react-custom-scrollbars": "^4.0.6",
     "@types/react-dates": "^17.1.7",
     "@types/react-responsive": "^8.0.2",
     "@types/reselect": "^2.2.0",
