+++ front/src/reducers/store.tsx
@@ -2,14 +2,14 @@ import { createStore, combineReducers, compose, applyMiddleware } from 'redux';
 // import PracticeCalendarReducer from './PracticeCalendarReducer';
 import {PracticeScheduleReducer} from './PracticeScheduleReducer'
 import {PracticeScheduleState} from '../types/PracticeScheduleState'
-import {PracticeCalendarMobileReducer } from './PracticeCalendarMobileReducer'
-import {PracticeCalendarMobileState} from '../types/PracticeCalendarMobileState'
+import {PracticeCalendarReducer } from './PracticeCalendarReducer'
+import {PracticeCalendarState} from '../types/PracticeCalendarState'
 
 // Redux-Thunk関連（非同期データ取得用）
 import thunk from 'redux-thunk'
 
 export type RootState = {
-  practiceCalendarMobile: PracticeCalendarMobileState;
+  practiceCalendar: PracticeCalendarState;
   practiceSchedule: PracticeScheduleState;
 };
 
@@ -17,7 +17,7 @@ export const store = createStore(
   combineReducers({
     // PracticeCalendarReducer: PracticeCalendarReducer,
     practiceSchedule: PracticeScheduleReducer,
-    practiceCalendarMobile: PracticeCalendarMobileReducer
+    practiceCalendar: PracticeCalendarReducer
   }),
   compose(
     applyMiddleware(thunk),
