--- front/src/selectors/PracticeCalendarMobileSelector.tsx
@@ -1,8 +1,0 @@
-import { createSelector } from 'reselect';
-import { State }  from '../reducers/store';
-import { PracticeCalendarMobileState } from '../types/PracticeCalendarMobileState';
-
-export const GetPracticeCalendarMobileState = createSelector(
-  (state: State) => state.practiceCalendarMobile,
-  (state: PracticeCalendarMobileState) => state
-);
