+++ front/src/selectors/PracticeCalendarSelector.tsx
@@ -1,0 +1,8 @@
+import { createSelector } from 'reselect';
+import { State }  from '../reducers/store';
+import { PracticeCalendarState } from '../types/PracticeCalendarState';
+
+export const GetPracticeCalendarState = createSelector(
+  (state: State) => state.practiceCalendar,
+  (state: PracticeCalendarState) => state
+);
