+++ front/src/types/PracticeCalendarState.tsx
@@ -1,7 +1,7 @@
 import moment from 'moment'
 
 // 練習カレンダー(mobile)のウィンドウ開閉
-export type PracticeCalendarMobileState = {
+export type PracticeCalendarState = {
     isOpen: boolean,
     date: moment.Moment | null,
     focused: boolean,
